requirements/base.txt
# coding: utf-8

# Django
Django>=2.0
djangorestframework
drf-nested-routers
django-cors-headers

# deploy
gunicorn
raven
Fabric3

# influxdb sdk 用于数据监控的
influxdb

ipython
pygments  # We'll be using this for the code highlighting
hotqueue
requests
bs4
redis
lxml
flake8

# plotly，用于作图
plotly

# 生成测试数据
faker

# django集成的websocket
# channels
# channels_redis

# celery
celery

# error
raven
